% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellScore.R
\name{CellScore}
\alias{CellScore}
\title{Title}
\usage{
CellScore(
  object = object,
  genelist = NULL,
  method = method,
  parallel.sz,
  abs.ranking = TRUE
)
}
\arguments{
\item{object}{seurat object}

\item{genelist}{genelist for calculating score}

\item{method}{Method for calculating score. Include "ssGSEA" , "AddModuleScore"}

\item{parallel.sz}{number of kernels for calculation}

\item{abs.ranking}{logical value. Default TRUE}
}
\description{
Title
}
