% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeScore.R
\name{DeScore}
\alias{DeScore}
\title{Title}
\usage{
DeScore(
  object = object,
  test.method = "wilcox.test",
  hide.ns = T,
  ctr.group = ctr.group,
  label = "p.signif",
  ScoreType = "ssGSEA"
)
}
\arguments{
\item{object}{seurat object}

\item{test.method}{method to be used for comparing means}

\item{hide.ns}{logical value. If TRUE, hide ns symbol when displaying significance levels.}

\item{ctr.group}{a list for comparison. list index 2 VS list index 1}

\item{label}{label type. Include "p.signif" , "p.format" (shows the formatted p value).}

\item{ScoreType}{The type of score you want to compare. Include "ssGSEA" , "AddModuleScore"}
}
\description{
Title
}
