% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectArtifactSamples.R
\name{detectArtifactSamples}
\alias{detectArtifactSamples}
\title{Detect Artifact Samples in Spatial Transcriptomics Data}
\usage{
detectArtifactSamples(
  spe,
  n_neighbors = 18,
  features = c("expr_chrM_ratio"),
  samples = "sample_id",
  log2 = FALSE,
  n_cores = 1,
  name = NULL
)
}
\arguments{
\item{spe}{A \code{SpatialExperiment} object containing the spatial transcriptomics data.}

\item{n_neighbors}{The number of neighbors to consider for local variance calculation. Default is 18.}

\item{features}{A character vector of column names in \code{colData} representing the features for which to calculate local variance.
Default is \code{c("expr_chrM_ratio")}.}

\item{samples}{The name of the column in \code{colData} that contains sample identifiers. Default is "sample_id".}

\item{log2}{A logical value indicating whether to apply a log2 transformation to the features before calculating local variance.
Default is FALSE.}

\item{n_cores}{The number of cores to use for parallel processing. Default is 1.}

\item{name}{An optional name for the column in \code{colData} where the local variance results will be stored. If NULL,
a default name will be generated based on the feature name and number of neighbors.}
}
\value{
A named list where each element corresponds to a sample and indicates whether the distribution
of local variance for the specified features is bimodal (TRUE) or not (FALSE). The names of the list
elements are the sample identifiers.
}
\description{
This function detects artifact samples in spatial transcriptomics data based on the bimodality of local variance
distributions. It leverages local variance calculations and the \code{is.bimodal} function from the LaplacesDemon package
to assess bimodality in the local variance of specified features across samples.
}
\examples{
bimodal_results <- detectArtifactSamples(spe,
                                         n_neighbors = 18,
                                         features = c("expr_chrM_ratio"),
                                         samples = "sample_id",
                                         log2 = FALSE,
                                         n_cores = 1,
                                         name = "local_variance")

}
