% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectSpots.R
\name{DetectSpots}
\alias{DetectSpots}
\title{Title}
\usage{
DetectSpots(
  meta = meta,
  tumor_name = tumor_name,
  nearby_name = nearby_name,
  group_var = group_var,
  celltype_var
)
}
\arguments{
\item{celltype_var}{}
}
\description{
Title
}
