% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLPFC_artifact.R
\docType{data}
\name{DLPFC_artifact}
\alias{DLPFC_artifact}
\title{human DLPFC dataset with a technical artifact (hangnail).}
\format{
An SpatialExperiment object.
}
\source{
\href{https://github.com/LieberInstitute/spatialDLPFC}{spatialLIBD}
}
\usage{
data(DLPFC_artifact)
}
\description{
The \code{DLPFC_artifact} dataset is a \code{SpatialExperiment} object containing a
single-sample subset of the human dorsolateral prefrontal cortex (DLPFC)
dataset from Hukki-Myers et al. 2023. This particular sample ("Br2743_ant") is
included to demonstrate the identification and removal of technical artifacts
within spatial transcriptomics data. The dataset serves as an example for
artifact detection using the 'SpotSweeper' workflow.
}
\examples{
data(DLPFC_artifact)
}
\references{
Hukki-Myers et al. (2023) bioRxiv
(\href{https://www.biorxiv.org/content/10.1101/2023.02.15.528722v1}{bioRxiv})
}
\keyword{datasets}
