% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_SPE.R
\name{example_SPE}
\alias{example_SPE}
\title{Spatial Experiment Data Fetcher}
\usage{
example_SPE(data = "DLPFC")
}
\arguments{
\item{data}{A character string specifying the type of data to fetch.
Default is "DLPFC", which fetches the spatialExperiment object
for Dorsolateral Prefrontal Cortex.}
}
\value{
A \code{SpatialExperiment} object containing the fetched data.
}
\description{
This function fetches spatial experiment data. It currently supports fetching data for
the DLPFC (Dorsolateral Prefrontal Cortex) brain region. The data is retrieved using
the \code{spatialLIBD} package.
}
\examples{
spe <- example_SPE()
spe <- example_SPE(data = "DLPFC")

}
