% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_compatibility.R
\name{getSpatialCoords}
\alias{getSpatialCoords}
\title{Get spatial coordinates from either SpatialExperiment or Seurat object}
\usage{
getSpatialCoords(x, image_id = NULL)
}
\arguments{
\item{x}{SpatialExperiment or Seurat object}

\item{image_id}{For Seurat objects, which image to use (default: first image)}
}
\value{
Matrix of spatial coordinates with rows as spots and columns as x,y coordinates
}
\description{
Get spatial coordinates from either SpatialExperiment or Seurat object
}
\examples{
library(STexampleData)
spe <- Visium_humanDLPFC()
coords <- getSpatialCoords(spe)
head(coords)
}
