% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{Heterogeneity}
\alias{Heterogeneity}
\title{Title}
\usage{
Heterogeneity(
  object = object,
  TumorName = TumorName,
  celltype_var = celltype_var
)
}
\arguments{
\item{object}{seurat object}

\item{TumorName}{the variable name of tumor in the spot type information}

\item{celltype_var}{spot type column names in meta.data}
}
\description{
Title
}
