% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexScorePlot.R
\name{HexScorePlot}
\alias{HexScorePlot}
\title{Title}
\usage{
HexScorePlot(
  object = object,
  type = type.by,
  color = NULL,
  align = "v",
  size = 0.8,
  alpha = 1,
  ncol = 1,
  legend = "right",
  starshape = 6,
  ScoreType = "ssGSEA",
  plot.image = TRUE,
  common.legend = FALSE,
  limits = NULL
)
}
\arguments{
\item{object}{seurat object}

\item{type}{cell type}

\item{legend}{legend position}

\item{ScoreType}{The type of score you want to show. Include "ssGSEA" , "AddModuleScore"}

\item{plot.image}{logical value.}

\item{common.legend}{logical value.}

\item{limits}{color range}
}
\description{
Title
}
