% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexSpatialPlot.R
\name{HexSpatialPlot}
\alias{HexSpatialPlot}
\title{Title}
\usage{
HexSpatialPlot(
  object = object,
  group.by = group.by,
  color = NULL,
  align = "v",
  size = 0.8,
  alpha = 1,
  ncol = 1,
  legend = "right",
  starshape = 6,
  plot.image = TRUE,
  common.legend = FALSE,
  limits = NULL
)
}
\arguments{
\item{limits}{}
}
\description{
Title
}
