% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HexSpatPlot.R
\name{HexSpatPlot}
\alias{HexSpatPlot}
\title{Title}
\usage{
HexSpatPlot(
  meta = meta,
  color = NULL,
  limits = NULL,
  img = img,
  size = size,
  alpha = alpha,
  group.by = group.by,
  starshape = starshape,
  plot.image = plot.image
)
}
\arguments{
\item{plot.image}{}
}
\description{
Title
}
