% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImmuneState.R
\name{ImmuneState}
\alias{ImmuneState}
\title{Title}
\usage{
ImmuneState(
  object = object,
  cancerType = cancerType,
  method = method,
  parallel.sz = 1
)
}
\arguments{
\item{parallel.sz}{}
}
\description{
Title
}
