% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_compatibility.R
\name{is_seurat}
\alias{is_seurat}
\title{SpotSweeper Seurat Compatibility Layer}
\usage{
is_seurat(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if object is Seurat
}
\description{
Functions to enable SpotSweeper compatibility with Seurat spatial objects
alongside SpatialExperiment objects.
Check if object is a Seurat object
}
