% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_compatibility.R
\name{is_spatial_experiment}
\alias{is_spatial_experiment}
\title{Check if object is a SpatialExperiment object}
\usage{
is_spatial_experiment(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical indicating if object is SpatialExperiment
}
\description{
Check if object is a SpatialExperiment object
}
