% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localVariance.R
\name{localVariance}
\alias{localVariance}
\title{localVariance Function}
\usage{
localVariance(
  spe,
  n_neighbors = 36,
  features = c("expr_chrM_ratio"),
  samples = "sample_id",
  log2 = FALSE,
  n_cores = 1,
  name = NULL
)
}
\arguments{
\item{spe}{SpatialExperiment object with the following columns in colData: sample_id, sum_umi, sum_gene}

\item{features}{Features to use for variance calculation}

\item{samples}{Column in colData to use for sample ID}

\item{log2}{Whether to log2 transform the features}

\item{n_cores}{Number of cores to use for parallelization}

\item{name}{Name of the new column to add to colData}

\item{n_neighbor}{Number of nearest neighbors to use for variance calculation}
}
\value{
SpatialExperiment object with feature variance added to colData
}
\description{
This function does calculates the local variance based on kNN.
}
\examples{
library(SpotSweeper)
library(spatialLIBD)

spe <- fetch_data(type = "spe")

spe <- localVariance(spe, n_neighbors=36, n_cores=4, name="local_mito_variance_k36")
}
