% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_z.R
\name{modified_z}
\alias{modified_z}
\title{Modified Z-score Function}
\usage{
modified_z(x, s = 1.4826)
}
\arguments{
\item{x}{A numeric vector for which outliers need to be detected.}

\item{s}{A constant used in the MAD calculation, default is 1.4826.}
}
\value{
A numeric vector of standardized values. Outliers typically have values far from zero.
}
\description{
This function calculates and returns standardized (z-score like) values based on the
Median Absolute Deviation (MAD) method. It is a robust method to detect outliers in
data that might not be normally distributed. The function is a replication of the outliers()
that was once available in the spatialEco package.
}
\examples{
data <- c(1, 2, 3, 4, 5, 100)
outliers(data)
}
