% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheCalculate.R
\name{NicheCalculate}
\alias{NicheCalculate}
\title{Title Calculate the mean and log2FC value for every niche}
\usage{
NicheCalculate(Ecotype_P = Ecotype_P)
}
\arguments{
\item{Ecotype_P}{}
}
\description{
Title Calculate the mean and log2FC value for every niche
}
