% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheCluster.R
\name{NicheCluster}
\alias{NicheCluster}
\title{Title subtype niche cluster}
\usage{
NicheCluster(
  object = object,
  centers = centers,
  nstart = 25,
  ScoreType = "ssGSEA",
  color = NULL
)
}
\arguments{
\item{color}{}
}
\description{
Title subtype niche cluster
}
