% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheCluster.R
\name{NicheKmeans}
\alias{NicheKmeans}
\title{Title Determine the number of subtypes}
\usage{
NicheKmeans(
  object = object,
  K.max = 10,
  nstart = 25,
  B = 50,
  ncol = 2,
  ScoreType = "ssGSEA"
)
}
\arguments{
\item{ScoreType}{}
}
\description{
Title Determine the number of subtypes
}
