% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rshinyAnnotation.R
\name{singleplot}
\alias{singleplot}
\title{Title}
\usage{
singleplot(
  object = object,
  group.by = group.by,
  ident = ident,
  metadata = metadata,
  color = NULL,
  size = 0.8,
  alpha = 1,
  shape = 19,
  celltype_var = celltype_var,
  plot.image = TRUE,
  common.legend = FALSE,
  limits = NULL
)
}
\arguments{
\item{object}{seurat object}

\item{limits}{}
}
\description{
Title
}
