% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpotChat.R
\name{SpotChat}
\alias{SpotChat}
\title{Title}
\usage{
SpotChat(
  object = object,
  subSample = subSample,
  cores = cores,
  json.path = json.path,
  min.cells = 10
)
}
\arguments{
\item{object}{seurat object}

\item{subSample}{sample name. Ensure that the sample name is in the 'orig.ident' column}

\item{cores}{Number of cores used}

\item{json.path}{Sample corresponding JSON file path}

\item{min.cells}{}
}
\description{
Title
}
