% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpotSweeper.R
\name{SpotSweeper}
\alias{SpotSweeper}
\title{Title}
\usage{
SpotSweeper(
  object = object,
  group_var = group_var,
  celltype_var = celltype_var,
  tumor_name = tumor_name,
  nearby_name = nearby_name
)
}
\arguments{
\item{object}{seurat object}

\item{group_var}{group column names in meta.data}

\item{celltype_var}{spot type column names in meta.data}

\item{tumor_name}{the variable name of tumor in the spot type information}

\item{nearby_name}{the variable name of the spot to be analyzed that is adjacent to the tumor spot}
}
\description{
Title
}
