% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findArtifacts.R
\name{findArtifacts}
\alias{findArtifacts}
\title{Identify and annotate artifacts in spatial transcriptomics data}
\usage{
findArtifacts(
  spe,
  mito_percent = "expr_chrM_ratio",
  mito_sum = "expr_chrM",
  samples = "sample_id",
  n_rings = 5,
  log = TRUE,
  name = "artifact",
  var_output = TRUE
)
}
\arguments{
\item{spe}{A SingleCellExperiment object.}

\item{mito_percent}{The column name representing the mitochondrial percent.
Default is 'expr_chrM_ratio'.}

\item{mito_sum}{The column name representing sum mitochondrial expression.
Default is 'expr_chrM'.}

\item{samples}{The column name representing sample IDs. Default is
'sample_id'.}

\item{n_rings}{The number of rings for local mito variance calculation.
Default is 5.}

\item{log}{Logical, indicating whether to log1p transform specified features.
Default is TRUE.}

\item{name}{Prefix for the local variance column names. Default is
'artifact'.}

\item{var_output}{Logical, indicating whether to include local variances in
the output. Default is TRUE.}
}
\value{
Returns the modified SingleCellExperiment object with artifact
annotations.
}
\description{
This function identifies and annotates potential artifacts in spatial
transcriptomics data. Artifacts are detected based on local mito variance,
and the results are added to the original
SpatialExperiment (sce) object.
}
\examples{
library(SpotSweeper)
library(SpatialExperiment)
library(escheR)

data(DLPFC_artifact)
spe <- DLPFC_artifact

# find artifacts
spe <- findArtifacts(spe,
    mito_percent = "expr_chrM_ratio",
    mito_sum = "expr_chrM",
    n_rings = 5,
    name = "artifact"
)

plotQC(spe, metric="expr_chrM_ratio", outliers="artifact")

}
\seealso{
\code{\link{localVariance}}
}
