% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findArtifacts.R
\name{findArtifacts}
\alias{findArtifacts}
\title{findArtifacts Function}
\usage{
findArtifacts(
  spe,
  mito_ratio = "expr_chrM_ratio",
  mito_expr = "expr_chrM",
  samples = "sample_id",
  n_rings = 5,
  n_cores = 1,
  log2 = TRUE,
  name = "artifact",
  var_output = TRUE
)
}
\value{
SpatialExperiment object with feature variance added to colData
}
\description{
This function uses the local variance of mitochondrial ratio to detect large artifacts
such as tissue hangnails.
}
\examples{
spe <- findArtifacts(spe,
                     mito_ratio="expr_chrM_ratio",
                     mito_expr="expr_chrM",
                     name="artifact"
                   )

}
