% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_compatibility.R
\name{getMetadata}
\alias{getMetadata}
\title{Get metadata from either SpatialExperiment or Seurat object}
\usage{
getMetadata(x)
}
\arguments{
\item{x}{SpatialExperiment or Seurat object}
}
\value{
Data.frame of metadata
}
\description{
Get metadata from either SpatialExperiment or Seurat object
}
\examples{
library(STexampleData)
spe <- Visium_humanDLPFC()
metadata <- getMetadata(spe)
colnames(metadata)
}
