% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localOutliers.R
\name{localOutliers}
\alias{localOutliers}
\title{localOutliers Function}
\usage{
localOutliers(spe, k = 15, threshold = 3)
}
\arguments{
\item{spe}{SpatialExperiment object with the following columns in colData: sample_id, sum_umi, sum_gene}

\item{k}{Number of nearest neighbors to use for outlier detection}

\item{threshold}{Threshold for outlier detection}
}
\value{
SpatialExperiment object with update z_umi_outlier columns
}
\description{
This function does detects local outliers based on k-nearest neighbors.
}
\examples{
localOutliers(spe_example, k=15, threshold=2)
}
