% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutliersPDF.R
\name{plotOutliersPDF}
\alias{plotOutliersPDF}
\title{Plot Outlier Metrics to PDF}
\usage{
plotOutliersPDF(
  spe,
  samples = "sample_id",
  metric = "detected",
  outliers = "local_outliers",
  low_color = "white",
  high_color = "black",
  stroke = 1,
  width = 5,
  height = 5,
  fname
)
}
\arguments{
\item{spe}{A SpatialExperiment object containing the data to be plotted.}

\item{metric}{A character string specifying the metric to be visualized
in the plot. This metric should be a column name in \code{rowData(spe)}.}

\item{outliers}{A character string specifying the column name in \code{rowData(spe)}
that indicates whether a data point is considered an outlier. Default is
"local_outliers".}

\item{low_color}{A character string indicating the color to be used for the low end
of the gradient scale. Default is "white".}

\item{high_color}{A character string indicating the color to be used for the high end
of the gradient scale. Default is "black".}

\item{stroke}{A numeric value specifying the border thickness for outlier
points. Default is 1.}

\item{width}{A numeric value indicating the width of the PDF's pages. Default
is 5.}

\item{height}{A numeric value indicating the height of the PDF's pages. Default
is 5.}

\item{fname}{A character string specifying the path and name of the output PDF file.}

\item{sample_id}{A character string specifying the column name in \code{colData(spe)}
that contains unique sample identifiers. Default is "sample_id".}
}
\value{
Invisible NULL. The function's primary effect is the creation of a PDF
file at the specified location.
}
\description{
This function generates a PDF file containing plots for each sample in the
SpatialExperiment object, highlighting outliers based on specified metrics.
Each plot visualizes outlier metrics for a single sample, allowing for
easy comparison and analysis across samples.
}
\examples{
plotOutliersPDF(spe, fname="outliers_visualization.pdf")

}
