% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_compatibility.R
\name{setMetadata}
\alias{setMetadata}
\title{Set metadata for either SpatialExperiment or Seurat object}
\usage{
setMetadata(x, metadata)
}
\arguments{
\item{x}{SpatialExperiment or Seurat object}

\item{metadata}{Data.frame of metadata to set}
}
\value{
Modified object with updated metadata
}
\description{
Set metadata for either SpatialExperiment or Seurat object
}
\examples{
library(STexampleData)
spe <- Visium_humanDLPFC()
metadata <- getMetadata(spe)
metadata$new_column <- 1
spe_updated <- setMetadata(spe, metadata)
"new_column" \%in\% colnames(getMetadata(spe_updated))
}
