% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_compatibility.R
\name{validateMetadataColumns}
\alias{validateMetadataColumns}
\title{Validate columns in spatial-object metadata}
\usage{
validateMetadataColumns(x, required_columns)
}
\arguments{
\item{x}{A \code{SpatialExperiment} or \code{Seurat} object.}

\item{required_columns}{Character vector of required metadata column names.}
}
\value{
\code{TRUE}, invisibly. An error is raised if columns are missing.
}
\description{
Validate columns in spatial-object metadata
}
\examples{
spe <- STexampleData::Visium_humanDLPFC()
validateMetadataColumns(spe, c("sample_id", "in_tissue"))
}
